feat(align): place the distance axis origin and generalise to the 2024 season (#32)
* feat(align): place the distance axis origin, and generalise to the season

Running a second circuit through the pipeline found that the distance axis had
been taking its origin from an accident. Distance zero was wherever the
reference lap's telemetry happened to open. At Suzuka that was 30 m past the
timing line, so every other lap had data to spare before grid point zero; at
Bahrain it was 2 m past, most laps opened 12 m further on, and the resampler
extrapolated their time channel across the first grid points -- a silent
-0.187 s on every reconstructed lap, all of it in sector 1, invisible at the
one circuit the pipeline was built on.

The origin is now placed rather than inherited: distance zero moves to the
latest telemetry opening in the session, so every lap has a real sample at or
before it. Bahrain's lap residual median goes -0.187 -> +0.007 s and it passes
every F010 check; Suzuka is unchanged to within a hundredth and slightly
tighter (lap p95 0.130 -> 0.117 s, closure p95 0.147 -> 0.157 s).

Two supporting changes came out of the same investigation. The projection seed
is searched near the start of the line instead of being pinned to vertex zero,
which was only correct while vertex zero was the lap start. And the line gains
a 60 m lead-in so a lap that opened before the seed lap did has somewhere real
to project; worth about 10 % of Suzuka's spread and neutral at Bahrain. A
lead-*out* was tried and rejected: it cost Suzuka 0.044 s of closure at p95,
because samples past the line's end are better served by the speed-integrated
extension that was already there.

Attribution measured rather than assumed -- placing the origin is the fix, the
other two are refinements. The spec's stated mechanism (an open reference line
mis-projecting samples) was wrong: the 20 m offsets it cited came from the
pinned seed, not the gap.

The circuit table gains the other 23 rounds of 2024, from Wikipedia's circuits
list cross-checked against the FIA figure F008 entered for Suzuka. A first
parse of that table returned Pescara's 22.8 km for Silverstone, so every entry
is also asserted against the laps actually driven at that circuit.

src/validate/season.py summarises every session's own F010 verdict into one
table, per session and per circuit -- the deliverable that turns "one session,
one circuit" into a statement about the method. The pipeline DAG gains a
summarise task after the load, and the availability sensor learns that
FastF1's rate limit is worth waiting out rather than permanent.

Races then found a second class of error, in F011 rather than F008: the
quality contracts described a qualifying session at a sea-level circuit, not
the sport. DRS codes 2, 3, 11, 13 and 15 all appear in race telemetry and each
blocked a race until it was added, so the enumeration is replaced by the check
it should always have been -- that the value is a plausible status byte.
Throttle reaches 104 %. Mexico City sits at 2,240 m and reports 784 hPa
against a floor of 800. A race leaves far more of the grid unshared with the
reference lap than qualifying does. And the feed emits occasional nonsense --
49 samples of one race report a gear of 72 -- so per-sample channel bounds now
carry a small glitch tolerance, which global CLAUDE.md 3.1 requires and which
a test asserts does not reach uniqueness, keys, invariants or completeness.

Tests: 837 offline passed. New: the lead-in and the defect it removes on a
synthetic Bahrain-shaped circuit (15), the circuit table against real laps
(21 + 2 data-marked), the season summary (21), and the rate limit.

* fix(quality): describe the sport, not one qualifying session

Running the 2024 season through the quality gate showed which of its
thresholds described the one session they were profiled on rather than
Formula 1. Each of these blocked a whole race until it was corrected, and each
correction is recorded in F011's spec with the session and count that forced
it.

- DRS: the enumeration of the seven codes Suzuka Q contained becomes a
  plausible-status-byte check, range(16). Codes 2, 3, 11, 13 and 15 all appear
  in race telemetry and each was added one at a time before it was clear the
  enumeration described the sample, not the channel.
- Pressure: floor 800 -> 700 hPa. Mexico City sits at 2,240 m and reports
  784 hPa; the old floor described sea level and refused the Mexican Grand
  Prix outright.
- delta_t nulls: 0.5 % -> 2 %. A race leaves far more of the grid unshared
  with the reference lap than qualifying does; the Dutch Grand Prix measured
  0.65 %.
- Per-sample channel bounds gain a 1e-3 glitch tolerance. The feed emits
  occasional nonsense -- 49 samples of Miami R report gear 72, two samples of
  British R project more than 100 m from the line -- and global CLAUDE.md 3.1
  says a corrupt record is skipped, not that the batch stops. A test asserts
  the tolerance reaches no Unique, ForeignKey or Invariant rule.
- n_gear: Int8 -> Int16 in the ingest and grid schemas. One sample of the
  Japanese Grand Prix reads 128, exactly one past int8, and the safe cast
  crashed the ingest. The warehouse column was already smallint; the value now
  flows through to be judged by the gate rather than killing the session.

The quality task no longer retries: a contract failure is a deterministic
property of the data just written, and three attempts cost twelve minutes per
failing session during the backfill for the same verdict.

F016's document is re-derived from the re-aligned Suzuka run and its test
follows the session's new slug. The four findings did not move -- T5 16/16 at
+0.122 s, the T8-T9 artefact, reconciliation to 0.0033 s -- but grid zero now
sits in a different place, so the straights row, the sector-1 reconciliation
row and the trace window's coordinates read differently and were regenerated
rather than assumed.

Not amended, deliberately: events.Invariant(event_boundaries_are_ordered)
fails at Spa-Francorchamps, Zandvoort and Lusail. That is F009's segmentation
and the gate is right to refuse it; those sessions stay recorded in the
per-circuit table rather than being let through.

Tests: 844 offline passed.

* test(report): follow the Suzuka session to its F006 slug and re-derive criterion 1

F016's acceptance test read the Suzuka run from the retired `2024_Japan_Q`
roots. The backfill re-ingested the session under the schedule's event name
(`2024_Japanese-Grand-Prix_Q_projection`) and F015 placed the axis origin,
so the corner phases hold to a thousandth while the straights row and the
total read from a grid zero further down the road: total +0.188 s (was
+0.081), straight +0.061 s (was -0.048). The expected figures are taken from
the regenerated summary.json, not typed.

Refs #31


* test: follow Suzuka to its live roots and re-derive the figures the placed origin moved

Seven test files guarded their Suzuka acceptance classes with
skipif(not PRESENT) on the retired `2026-09-01/2024_Japan_Q` and
`2024_Japan_Q_projection` roots. When the backfill re-ingested Suzuka under
the schedule's event name the guards turned those classes into silent skips
and the `-m data` count hid it. Re-pointed to
`2026-09-05/2024_Japanese-Grand-Prix_Q` and `*_Q_projection`; the four
assertions that then failed carried figures F015's origin placement moved,
re-derived with a note: 57 fixed 100 m bins (was 58), 92 micro-sectors per
lap and 6,808 rows (was 93 / 6,882), warehouse grid rows 42,198 (was 42,418
-- 74 laps x 3 grid points). Lap count 74 and 592 corner metrics unchanged.
The seven files run 117 tests with no skips; the offline suite exits 0.

Refs #31


---------

+++ src/grid/resample.py
@@ -47,7 +47,9 @@ GRID_SCHEMA: dict[str, str] = {
     "rpm": "float32",
     "x": "float32",
     "y": "float32",
-    "n_gear": "Int8",
+    # Int16 for the same reason as the ingest schema: one sample of the 2024
+    # Japanese Grand Prix reads gear 128, one past int8 (F015).
+    "n_gear": "Int16",
     "drs": "Int8",
     "brake": "boolean",
     "source_gap_m": "float32",
